+++ app/src/main/java/my/life/tracker/finanzas/interfaces/RecordDao.kt
@@ -13,7 +13,7 @@ interface RecordDao {
     fun deleteRecord(record:Record)
 
     @Query("SELECT * FROM Record WHERE Record.id = :id")
-    fun getRecordById(id: Int) : Record
+    fun getRecordById(id: Long) : Record
 
     @Query("SELECT * FROM Record WHERE Record.date BETWEEN :from AND :to ORDER BY Record.date DESC")
     fun getRecords(from: String, to: String) : List<Record>
