+++ app/src/main/java/my/life/tracker/finanzas/interfaces/TagDao.kt
@@ -9,16 +9,16 @@ interface TagDao {
     @Upsert
     fun addTag(tag: Tag)
     @Query("SELECT * FROM Tag WHERE id = :id")
-    fun getTagById(id:Int): Tag?
+    fun getTagById(id:Long): Tag?
 
     @Delete
     fun deleteTag(tag: Tag)
     @Query("delete from Tag where id in (:idList)")
-    fun deleteTags(idList: List<Int>)
+    fun deleteTags(idList: List<Long>)
 
     @Query("SELECT * FROM Tag")
     fun getTags() : List<Tag>
 
     @Query("SELECT * FROM Tag WHERE Tag.id IN (:tags)")
-    fun getTagsByIdList(tags:List<Int>) : List<Tag>
+    fun getTagsByIdList(tags:List<Long>) : List<Tag>
 }
