+++ app/src/main/java/my/life/tracker/finanzas/interfaces/TagRecordDao.kt
@@ -13,19 +13,19 @@ interface TagRecordDao {
     fun deleteTagRecord(tag: TagRecord)
 
     @Query("delete from TagRecord where id in (:idList)")
-    fun deleteTagRecords(idList: List<Int>)
+    fun deleteTagRecords(idList: List<Long>)
 
     @Query("SELECT * FROM TagRecord")
     fun getTagRecords() : List<TagRecord>
 
     @Query("SELECT * FROM TagRecord WHERE TagRecord.idRecord = :id")
-    fun getAllTagRecordByRecordId(id:Int) : List<TagRecord>
+    fun getAllTagRecordByRecordId(id:Long) : List<TagRecord>
 
     @Query("SELECT * FROM TagRecord WHERE TagRecord.idTag = :id")
-    fun getAllTagRecordByTagId(id:Int) : List<TagRecord>
+    fun getAllTagRecordByTagId(id:Long) : List<TagRecord>
 
     @Query("SELECT * FROM TagRecord WHERE TagRecord.idRecord in (:idList)")
-    fun getAllTagRecordByRecordIdList(idList:List<Int>) : List<TagRecord>
+    fun getAllTagRecordByRecordIdList(idList:List<Long>) : List<TagRecord>
 
 
 }
